fix(protocol): Defaults.NativeType should conform to Defaults.Serializable
