+++ Sources/Defaults/Defaults+Protocol.swift
@@ -39,33 +39,3 @@ public protocol DefaultsBridge {
 
 /// Convenience protocol for `Codable`
 protocol DefaultsCodableBridge: DefaultsBridge where Serializable == String, Value: Codable {}
-
-/**
-`NativeForm` is a type that we want it to store in the `UserDefaults`
-It should have a associated type named `CodableForm` which protocol conform to `Codable` and `Defaults.Serializable`
-So we can convert the json string into `NativeType` like this.
-```
-guard
-	let jsonString = string,
-	let jsonData = jsonString.data(using: .utf8),
-	let codable = try? JSONDecoder().decode(NativeType.CodableForm.self, from: jsonData)
-else {
-	return nil
-}
-
-return codable.toNative()
-```
-*/
-public protocol DefaultsNativeType {
-	associatedtype CodableForm: Defaults.Serializable, Defaults.CodableType
-}
-
-/**
-CodableType is a type that stored in the `UserDefaults` previously, now needs to be migrated.
-It should have an associated type name `NativeForm` which is the type we want it to store in `UserDefaults`.
-And it also have a `toNative()` function to convert itself into `NativeForm`.
-*/
-public protocol DefaultsCodableType: Codable {
-	associatedtype NativeForm: Defaults.Serializable, Defaults.NativeType
-	func toNative() -> NativeForm
-}
