+++ Sources/Defaults/UserDefaults.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 extension UserDefaults {
-	private func _get<Value: Defaults.Serializable>(_ key: String) -> Value? {
+	func _get<Value: Defaults.Serializable>(_ key: String) -> Value? {
 		let anyObject = object(forKey: key)
 
 		/// Return directly if anyObject can cast to Value, means `Value` is Native supported type.
@@ -14,7 +14,7 @@ extension UserDefaults {
 		return nil
 	}
 
-	private func _set<Value: Defaults.Serializable>(_ key: String, to value: Value) {
+	func _set<Value: Defaults.Serializable>(_ key: String, to value: Value) {
 		if (value as? _DefaultsOptionalType)?.isNil == true {
 			removeObject(forKey: key)
 			return
@@ -25,32 +25,6 @@ extension UserDefaults {
 		} else if let serialized = Value.bridge.serialize(value as? Value.Value) {
 			set(serialized, forKey: key)
 		}
-	}
-
-	/**
-	Get json string in `UserDefaults` and decode it into the `NativeForm`.
-
-	How it works?
-	For example:
-	Step1. If `Value` is  `[String]`, `Value.CodableForm` will covert into `[String].CodableForm`.
-	`JSONDecoder().decode([String].CodableForm.self, from: jsonData)`
-
-	Step2. `Array`conform to `NativeType`, it's `CodableForm` is `[Element.CodableForm]` and `Element` is `String`.
-	`JSONDecoder().decode([String.CodableForm].self, from: jsonData)`
-
-	Step3. `String`'s `CodableForm` is `self`,  because `String` is `Codable`.
-	`JSONDecoder().decode([String].self, from: jsonData)`
-	*/
-	func migration<Value: Defaults.Serializable & Defaults.NativeType>(forKey key: String, of type: Value.Type) {
-		guard
-			let jsonString = string(forKey: key),
-			let jsonData = jsonString.data(using: .utf8),
-			let codable = try? JSONDecoder().decode(Value.CodableForm.self, from: jsonData)
-		else {
-			return
-		}
-
-		_set(key, to: codable.toNative())
 	}
 
 	public subscript<Value: Defaults.Serializable>(key: Defaults.Key<Value>) -> Value {
